Rename and fix imports
